Move fluent client generators
